+++ src/test/java/io/beanmapper/testmodel/beanproperty/SourceBeanProperty.java
@@ -1,0 +1,10 @@
+package io.beanmapper.testmodel.beanproperty;
+
+import io.beanmapper.annotations.BeanProperty;
+
+public class SourceBeanProperty {
+
+    @BeanProperty(name = "age2")
+    public Integer age;
+
+}