+++ src/test/java/io/beanmapper/testmodel/beanproperty/SourceNestedBeanProperty.java
@@ -1,0 +1,13 @@
+package io.beanmapper.testmodel.beanproperty;
+
+import io.beanmapper.annotations.BeanProperty;
+
+public class SourceNestedBeanProperty {
+
+    @BeanProperty(name = "alpha.beta.gamma")
+    public String value1;
+
+    @BeanProperty(name = "alpha.beta.delta")
+    public String value2;
+
+}