+++ src/test/java/io/beanmapper/testmodel/beanproperty/TargetNestedBeanProperty.java
@@ -1,0 +1,6 @@
+package io.beanmapper.testmodel.beanproperty;
+
+public class TargetNestedBeanProperty {
+
+    public TargetNestedBeanPropertyAlpha alpha;
+}