+++ src/test/java/io/beanmapper/testmodel/beanproperty/TargetNestedBeanPropertyAlpha.java
@@ -1,0 +1,6 @@
+package io.beanmapper.testmodel.beanproperty;
+
+public class TargetNestedBeanPropertyAlpha {
+
+    public TargetNestedBeanPropertyBeta beta;
+}