+++ src/test/java/io/beanmapper/testmodel/beanproperty/TargetNestedBeanPropertyBeta.java
@@ -1,0 +1,6 @@
+package io.beanmapper.testmodel.beanproperty;
+
+public class TargetNestedBeanPropertyBeta {
+
+    public Long gamma;
+}